requirements.txt
PyYAML>=5.4
boto3
numpy
tweepy